+++ src/lib/include/ml_log.hpp
@@ -21,15 +21,15 @@ public:
 
 	/* Compute the cost of the provided parameters for the provided data set */
 	static Matrix *sigmoid(Matrix &z);
-	static double computeCost(Matrix &training_X, Matrix &training_y, Matrix &training_theta);
+	static float computeCost(Matrix &training_X, Matrix &training_y, Matrix &training_theta);
 	static Matrix *gradientCalculate(Matrix &training_X, Matrix &training_y, Matrix &theta);
-	static Matrix *GradientDescent(Matrix &training_X, Matrix &training_y, Matrix &theta, double alpha, int num_iterations);
-	static Matrix *Predict(Matrix &input_examples, Matrix &theta, double threshold);
+	static Matrix *GradientDescent(Matrix &training_X, Matrix &training_y, Matrix &theta, float alpha, int num_iterations);
+	static Matrix *Predict(Matrix &input_examples, Matrix &theta, float threshold);
 
 	/* Regularized Operations */
-	static double computeCost(Matrix &training_X, Matrix &training_y, Matrix &training_theta, double regularizationParam);
-	static Matrix *gradientCalculate(Matrix &training_X, Matrix &training_y, Matrix &theta, double regularizationParam);
-	static Matrix *GradientDescent(Matrix &training_X, Matrix &training_y, Matrix &theta, double alpha, int num_iterations, double regularizationParam);
+	static float computeCost(Matrix &training_X, Matrix &training_y, Matrix &training_theta, float regularizationParam);
+	static Matrix *gradientCalculate(Matrix &training_X, Matrix &training_y, Matrix &theta, float regularizationParam);
+	static Matrix *GradientDescent(Matrix &training_X, Matrix &training_y, Matrix &theta, float alpha, int num_iterations, float regularizationParam);
 
 	/* Multi-Class Classification */
 	static Matrix *OneVsAll(Matrix &training_X, Matrix &training_y, int num_classes, double alpha, int num_iterations, double regularizationParam);
